requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
openai>=1.30.0
azure-storage-blob
azure-ai-documentintelligence
azure-cosmos
azure-identity
azure-core>=1.29.0